Brand new bugs! Fixed them:
1) When removing files in the purged user's home dir, os.path.isdir() was being
called on the files, but not their absolute paths, to determine if they were
directories. This is not going to have predictable results, so it will not
always be possible to remove subdirectories of the user's home directory.  Fixed
by testing with os.path.isdir() on the absolute path name of the
files/directories. 

2) archiveFile was getting set to /home/foo/.tar.gz instead of /home/foo.tar.gz.
Changed a comma to a plus to fix that one.